chore: BaseViewModelV2 -> BaseViewModel 이름 변경
