+++ android/presentation/src/main/java/com/hmju/presentation/base/FragmentViewModel.kt
@@ -5,6 +5,6 @@ package com.hmju.presentation.base
  * Fragment 사용되는 공통 LiveData 들을 정의합니다.
  * Created by juhongmin on 2022/04/15
  */
-open class FragmentViewModel : BaseViewModelV2() {
+open class FragmentViewModel : BaseViewModel() {
 
 }
